refactor: extract post confirmation service
